[ BE ] (perf) ItemService 인터페이스 updateItem() 메서드 매개변수 수정
[ BE ] (perf) ItemService 클래스 updateItem() 메서드 매개변수 수정

상품 수량의 타입을 int 로 수정해주었습니다.

+++ 4 WEEK/item-service/src/main/java/hello/itemservice/item/ItemService.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Service;
 
 public interface ItemService {
     void join(Item item);
-    void updateItem(Item item, String itemName, double itemPrice, long itemQuantity);
+    void updateItem(Item item, String itemName, double itemPrice, int itemQuantity);
     Item findItem(Long itemId);
 }
